Closes sr #2208: Merge of r5187 from /branches/poker-prizes@r5187 to trunk
This change adds a new implementation for PokerPrizes, which is a set of
classes currently used only by PokerTournament.  The long term goal is to
allow prize structures to be looked up without necessarily having an
existing PokerTournament instance.

Mostly, this implementation ripped out the prizes portions of
PokerTournament as the starting point for the new class.

git-svn-id: http://svn.gna.org/svn/pokersource/trunk@5198 c4f34931-41c7-4605-b615-9bb9f479c57e